Update manual_example_multiple_serial_jobs.md
Slurm on COSMOS behaves differently form slum on Aurora. The srun option --overlap has been removed from the example, because it is not needed for a single node on Aurora (as in the example) and is detrimental on COSMOS. The tip has also been updated. Otherwise, everything else still refers to Aurora and needs to be updated when COSMOS opens