Bad Aurora instructions removed
On Aurora srun in the taskfarm required --exclusive --overlap in that order to distribute tasks over multiple nodes, which was probably a bug, because the two instructions are mutually exclusive on Cosmos. --overlap, which is not needed, has been removed in the example script and so has the reference to --exclusive --overlap in the tip.